requirements.txt
Flask
psycopg2-binary
requests
pyopenssl
discord
